Add Manual Override of Outputs on Dash/Flex Display
Added the ability to manually override outputs (igniter, auger, fan) on the default dashboard and on the flex touch display.  The user can enable this in the safety settings, and select the amount of seconds to override the controller during operation.  Use with caution as this can potentially leave the igniter on for extended periods of time.