Helpful Server Error Page
Added a more helpful server error page that is loaded by nginx if the server is experiencing errors.  Attempts a reload of the root/dash every 10 seconds.  Links provided to resources such as the admin page, the Supervisor page (if setup), documentation and help sites.

This page will only be configured on new installations.