UV for New Installations and PIP List in Admin Page
Changed install script to utilize UV for creating a VENV and installing dependencies.  This results in a much faster installation process, quicker upgrades and street cred.

Quality of Life: Added Python Module List to the admin page (Admin > Module List).  This will show you the modules that are installed and their version numbers.  This is useful to debug common package version issues.